Moved the accept/complete quest class into the QuestManager and added docstrings. Modified DisplayDialogue according to the changes but probably needs changes regarding player etc.
