--- src/main/java/quest_system/AcceptQuest.java
@@ -1,11 +1,0 @@
-package quest_system;
-
-public class AcceptQuest {
-
-    public void accept(Quest quest) {
-        assert !quest.isAccepted();
-
-        quest.toggleAccepted();
-    }
-
-}
