--- src/main/java/quest_system/CompleteQuest.java
@@ -1,20 +1,0 @@
-package quest_system;
-
-import characters.GameCharacter;
-import quest_system.Quest;
-
-public class CompleteQuest {
-
-    public void complete(GameCharacter player, Quest quest) {
-        assert !quest.isComplete();
-
-        for (String item : quest.getRewards()) {
-            player.addItem(item);
-        }
-
-        quest.toggleComplete();
-
-
-    }
-
-}
