+++ src/main/java/scene_system/DisplayDialogue.java
@@ -1,8 +1,7 @@
 package scene_system;
 
 import characters.PlayerCharacter;
-import quest_system.AcceptQuest;
-import quest_system.CompleteQuest;
+import quest_system.QuestManager;
 import quest_system.Quest;
 import characters.*;
 
@@ -11,19 +10,17 @@ public class DisplayDialogue {
 
     public String dialogue(NonPlayerCharacter chara, PlayerCharacter player){
         Quest quest = chara.getQuest();
-        AcceptQuest acceptQuest = new AcceptQuest();
-        CompleteQuest completeQuest = new CompleteQuest();
         if(quest == null){
             return "ew stinky";
         }
         switch (quest.getCompletion()) {
             case 0:
-                acceptQuest.accept(quest);
+                player.accept(quest);
                 return chara.getQuestDialogue(chara.BEGIN_QUEST);
 
             case 1:
                 if (quest.checkDone(player)) {
-                    completeQuest.complete(player, quest);
+                    player.complete(player, quest);
                     return chara.getQuestDialogue(chara.END_QUEST);
                 }
                 return chara.getQuestDialogue(chara.DURING_QUEST);
